+++ src/clients/identity/exceptions/max-age-exceeded.exception.ts
@@ -1,0 +1,7 @@
+import { StatusCodeError } from './status-code-error.exception';
+
+export class MaxAgeExceededException extends StatusCodeError {
+  constructor() {
+    super(401, 'Max age exceeded');
+  }
+}