+++ src/clients/identity/step-up/types.ts
@@ -1,0 +1,9 @@
+export interface ValidateStepupFields {
+  amr?: string[];
+  acr?: string;
+  auth_time?: number;
+}
+
+export interface IValidateStepupTokenOptions {
+  maxAge?: number;
+}